Simplify message sending.
This had become quite a mess. I had assumed initially that SOM had continuations
but hadn't factored in the `restart bytecode. The former assumption made things
a too generic (including boxing return values, which is clearly inefficient);
the latter (mis)assumption made things messy. This commit goes some way to
fixing things up.

The basic idea here is to find a sensible way of dealing with closures returning
from a parent function. The basic solution here is to introduce a new `enum`
`SendReturn` which captures the idea that functions can return normal values,
errors, or percolate a return value up the call stack. As a happy bonus, this
allows us to easily place return values (whether "normal" or "closure" returns)
at the right place on the stack so we no longer need to box return values.

So as well as making the code easier to read, this is also a little faster: 5%
on a (silly, small) `while` benchmark.